docs(build) deploy to angular-ui/gh-pages

+++ deploy-docs.sh
@@ -1,17 +1,13 @@
 #!/bin/bash
-set -e # exit with nonzero exit code if anything fails
+set -e 
 
-echo ${TRAVIS_SECURE_ENV_VARS}
+[[ $TRAVIS_SECURE_ENV_VARS == "true" ]] || { echo "No github key avaliable, aborting publishing"; exit 0; }
 
-COMMIT_ID="$(git rev-parse --short HEAD)"
+ID_REF="$(git rev-parse --short HEAD)"
 
-mkdir docs-out;
+git clone "https://${GH_KEY}@${GH_REF}" ./docs-out -b ${GH_PAGES_BRANCH} --single-branch --depth=1
 
-# go to the out directory and create a *new* Git repo
 cd docs-out
-git clone "https://${GH_KEY}@${GH_REF}" . -b ${GH_PAGES_BRANCH} --single-branch --depth=1
-
-git checkout ${GH_PAGES_BRANCH}
 
 # clear out everything
 git rm -rf .
@@ -28,11 +24,8 @@ git config user.email "travisci@users.noreply.github.com"
 
 # The first and only commit to this new Git repo contains all the
 # files present with the commit message "Deploy to GitHub Pages".
-git commit -m "docs(all) new deployment (ref: angular-ui/ui-select@${COMMIT_ID})"
+git commit -m "docs(*): new deploy (angular-ui/ui-select@${ID_REF})"
 
-# Force push from the current repo's master branch to the remote
-# repo's gh-pages branch. (All previous history on the gh-pages branch
-# will be lost, since we are overwriting it.) We redirect any output to
-# /dev/null to hide any sensitive credential data that might otherwise be exposed.
+
 git push origin --quiet 
-#"https://${GH_KEY}@${GH_REF}" master:gh-pages2 > /dev/null 2>&1
+#> /dev/null 2>&1
